threads-worker/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
celery==5.4.0